% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_full_report.R
\name{build_complete_html_report}
\alias{build_complete_html_report}
\title{Build complete HTML report with all sections}
\usage{
build_complete_html_report(
  metrics,
  dqd_data,
  dqd_scores,
  group_dqd_scores,
  table_dqd_scores,
  table_groups,
  num_participants,
  total_rows_removed,
  type_concept_summary,
  has_delivery_data = TRUE,
  has_dqd_data = TRUE
)
}
\description{
Build complete HTML report with all sections
}
