% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_data_unavailable_message}
\alias{build_data_unavailable_message}
\title{Build data unavailable message box}
\usage{
build_data_unavailable_message(data_type, section_title)
}
\arguments{
\item{data_type}{Character. Type of data that's unavailable (e.g., "DQD", "Delivery")}

\item{section_title}{Character. Title of the section}
}
\value{
HTML string with warning message
}
\description{
Build data unavailable message box
}
