% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_delivery_overview_section}
\alias{build_delivery_overview_section}
\title{Build delivery overview section}
\usage{
build_delivery_overview_section(
  metrics,
  dqd_scores,
  num_participants,
  total_rows_removed,
  type_concept_summary,
  has_delivery_data = TRUE,
  has_dqd_data = TRUE
)
}
\description{
Build delivery overview section
}
