% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_delivery_report_section}
\alias{build_delivery_report_section}
\title{Build delivery report section with table groups}
\usage{
build_delivery_report_section(
  metrics,
  table_groups,
  group_dqd_scores,
  has_delivery_data = TRUE,
  has_dqd_data = TRUE
)
}
\description{
Build delivery report section with table groups
}
