% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builder.R
\name{build_delivery_report_section_for_report}
\alias{build_delivery_report_section_for_report}
\title{Build delivery report section (uses html_builders.R)}
\usage{
build_delivery_report_section_for_report(
  metrics,
  table_groups,
  group_dqd_scores,
  table_dqd_scores,
  num_participants,
  has_delivery_data
)
}
\description{
Build delivery report section (uses html_builders.R)
}
