% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R, R/report_builder.R
\name{build_dqd_grid_section}
\alias{build_dqd_grid_section}
\title{Build DQD grid section}
\usage{
build_dqd_grid_section(dqd_scores, has_dqd_data)

build_dqd_grid_section(dqd_scores, has_dqd_data)
}
\arguments{
\item{dqd_scores}{List with DQD scores and grid}

\item{has_dqd_data}{Logical}
}
\value{
Character HTML string
}
\description{
Generates the DQD results grid table.
}
