% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_report_header}
\alias{build_report_header}
\title{Build report header}
\usage{
build_report_header(
  metrics,
  dqd_score,
  has_delivery_data = TRUE,
  has_dqd_data = TRUE
)
}
\description{
Build report header
}
