% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_sidebar}
\alias{build_sidebar}
\title{Build sidebar navigation}
\usage{
build_sidebar(
  metrics,
  dqd_score,
  has_delivery_data = TRUE,
  has_dqd_data = TRUE
)
}
\description{
Build sidebar navigation
}
