% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_table_drilldown_section}
\alias{build_table_drilldown_section}
\title{Build table drilldown section (initially hidden)}
\usage{
build_table_drilldown_section()
}
\description{
Build table drilldown section (initially hidden)
}
