% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_table_group_content}
\alias{build_table_group_content}
\title{Build content for a single table group}
\usage{
build_table_group_content(group_name, group_tables, metrics, group_dqd_score)
}
\description{
Build content for a single table group
}
