% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_builders.R
\name{build_table_rows_data}
\alias{build_table_rows_data}
\title{Build table rows JSON for a table group}
\usage{
build_table_rows_data(group_tables, all_table_data, num_participants)
}
\arguments{
\item{group_tables}{Character vector of table names}

\item{all_table_data}{List of table data structures}

\item{num_participants}{Integer total participants}
}
\value{
List of row data (will be serialized to JSON)
}
\description{
Prepares JSON array of table row data for the delivery report table.
Each row contains the metrics displayed in the main table grid.
}
