% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_time_series_section}
\alias{build_time_series_section}
\title{Build time series section}
\usage{
build_time_series_section(metrics, has_delivery_data = TRUE)
}
\description{
Build time series section
}
