% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builder.R
\name{build_time_series_section_for_report}
\alias{build_time_series_section_for_report}
\title{Build time series section (uses html_builders.R)}
\usage{
build_time_series_section_for_report(metrics, has_delivery_data)
}
\description{
Build time series section (uses html_builders.R)
}
