% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builder.R
\name{build_vocab_harmonization_section_for_report}
\alias{build_vocab_harmonization_section_for_report}
\title{Build vocab harmonization section (uses html_builders.R)}
\usage{
build_vocab_harmonization_section_for_report(metrics, has_delivery_data)
}
\description{
Build vocab harmonization section (uses html_builders.R)
}
