% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_builders.R
\name{build_vocabulary_harmonization_section}
\alias{build_vocabulary_harmonization_section}
\title{Build vocabulary harmonization section}
\usage{
build_vocabulary_harmonization_section(metrics, has_delivery_data = TRUE)
}
\description{
Build vocabulary harmonization section
}
