% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{calculate_table_dqd_score}
\alias{calculate_table_dqd_score}
\title{Calculate DQD score for a single table}
\usage{
calculate_table_dqd_score(dqd_data, table_name)
}
\arguments{
\item{dqd_data}{Tibble from load_dqd_results()}

\item{table_name}{Character string of table name}
}
\value{
Numeric percentage (0-100)
}
\description{
Calculate DQD score for a single table
}
