% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{calculate_table_group_dqd_score}
\alias{calculate_table_group_dqd_score}
\title{Calculate DQD score for specific tables}
\usage{
calculate_table_group_dqd_score(dqd_data, tables)
}
\arguments{
\item{dqd_data}{Tibble from load_dqd_results()}

\item{tables}{Character vector of table names}
}
\value{
Numeric percentage (0-100)
}
\description{
Calculate DQD score for specific tables
}
