% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformers.R
\name{calculate_total_rows_removed}
\alias{calculate_total_rows_removed}
\title{Calculate total rows removed due to missing person_id}
\usage{
calculate_total_rows_removed(metrics)
}
\arguments{
\item{metrics}{List of metric data frames}
}
\value{
Integer total rows removed
}
\description{
Calculate total rows removed due to missing person_id
}
