% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformers.R
\name{create_empty_metrics}
\alias{create_empty_metrics}
\title{Create empty metrics structure}
\usage{
create_empty_metrics()
}
\value{
List with empty metric structures
}
\description{
Used when delivery data is not available.
}
