% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_calculators.R
\name{format_number}
\alias{format_number}
\title{Format numeric value for display}
\usage{
format_number(value, big_mark = ",")
}
\arguments{
\item{value}{Numeric value}

\item{big_mark}{Character to use as thousands separator (default: ",")}
}
\value{
Character formatted number
}
\description{
Format numeric value for display
}
