% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builder.R
\name{generate_dqd_grid_rows}
\alias{generate_dqd_grid_rows}
\title{Generate DQD grid table rows}
\usage{
generate_dqd_grid_rows(grid)
}
\arguments{
\item{grid}{Data frame with DQD grid data}
}
\value{
Character HTML string with table rows
}
\description{
Generate DQD grid table rows
}
