% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_table_drilldown_data}
\alias{generate_table_drilldown_data}
\title{Generate table drilldown data}
\usage{
generate_table_drilldown_data(table_name, metrics, dqd_data)
}
\arguments{
\item{table_name}{Name of the table}

\item{metrics}{Parsed metrics}

\item{dqd_data}{DQD results data}
}
\value{
List with all drilldown information
}
\description{
Generate table drilldown data
}
