% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_default_dates_threshold}
\alias{get_default_dates_threshold}
\title{Get alert threshold for default dates}
\usage{
get_default_dates_threshold(config = default_config())
}
\arguments{
\item{config}{Configuration list from merge_config()}
}
\value{
Numeric threshold percentage
}
\description{
Get alert threshold for default dates
}
