% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_dqd_thresholds}
\alias{get_dqd_thresholds}
\title{Get DQD score classification thresholds}
\usage{
get_dqd_thresholds(config = default_config())
}
\arguments{
\item{config}{Configuration list from merge_config()}
}
\value{
List with good_threshold and fair_threshold
}
\description{
Get DQD score classification thresholds
}
