% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{get_extended_palette}
\alias{get_extended_palette}
\title{Get extended color palette for individual type concepts}
\usage{
get_extended_palette(n)
}
\arguments{
\item{n}{Number of colors needed}
}
\value{
Vector of hex colors
}
\description{
Get extended color palette for individual type concepts
}
