% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_invalid_concepts_threshold}
\alias{get_invalid_concepts_threshold}
\title{Get alert threshold for invalid concepts}
\usage{
get_invalid_concepts_threshold(config = default_config())
}
\arguments{
\item{config}{Configuration list from merge_config()}
}
\value{
Numeric threshold count
}
\description{
Get alert threshold for invalid concepts
}
