% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_processors.R
\name{get_table_groups}
\alias{get_table_groups}
\title{Get table groups from configuration}
\usage{
get_table_groups(config = default_config())
}
\arguments{
\item{config}{Configuration list (default: default_config())}
}
\value{
Named list of table groups
}
\description{
Get table groups from configuration
}
