% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_type_concept_group_order}
\alias{get_type_concept_group_order}
\title{Get canonical order for type concept groups}
\usage{
get_type_concept_group_order(config = default_config())
}
\arguments{
\item{config}{Configuration list from merge_config()}
}
\value{
Character vector of group names in display order
}
\description{
Get canonical order for type concept groups
}
