% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{merge_config}
\alias{merge_config}
\title{Merge user configuration with defaults}
\usage{
merge_config(user_config = list())
}
\arguments{
\item{user_config}{List of user-specified configuration options}
}
\value{
Merged configuration list
}
\description{
Recursively merges user-specified configuration with default values.
User values override defaults.
}
\examples{
user_cfg <- list(thresholds = list(default_dates_pct = 5))
cfg <- merge_config(user_cfg)
}
