% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformers.R
\name{prepare_overall_transitions}
\alias{prepare_overall_transitions}
\title{Prepare overall transition statistics (dataset-wide)}
\usage{
prepare_overall_transitions(metrics)
}
\arguments{
\item{metrics}{List of metric data frames}
}
\value{
List with transition statistics
}
\description{
Prepare overall transition statistics (dataset-wide)
}
