% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformers.R
\name{prepare_overall_type_concepts}
\alias{prepare_overall_type_concepts}
\title{Prepare overall type concept summary (dataset-wide)}
\usage{
prepare_overall_type_concepts(metrics)
}
\arguments{
\item{metrics}{List of metric data frames}
}
\value{
Data frame with type concept summary
}
\description{
Prepare overall type concept summary (dataset-wide)
}
