% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_full_report.R
\name{prepare_report_data_json}
\alias{prepare_report_data_json}
\title{Prepare all report data as JSON for JavaScript access}
\usage{
prepare_report_data_json(
  metrics,
  dqd_data,
  table_groups,
  group_dqd_scores,
  table_dqd_scores
)
}
\description{
Prepare all report data as JSON for JavaScript access
}
