% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformers.R
\name{prepare_all_table_data}
\alias{prepare_all_table_data}
\title{Prepare table data for all tables}
\usage{
prepare_all_table_data(metrics, table_dqd_scores)
}
\arguments{
\item{metrics}{List of metric data frames}

\item{table_dqd_scores}{Named list of DQD scores per table}
}
\value{
List of table data structures
}
\description{
Prepare table data for all tables
}
